Update repo for stable/ocata
There are no new tags for the 3rd party repositories, so for now we will
have to use master branch. This should be fixed once we can get tags
or branches created for these repositories.

Additionally, the mirror that is used by default in the upstream
ansible-kafka role is not currently working, we can overwrite this with
a var in the meta/main.yml file.

Change-Id: Ia7dfb8f87bd65ccc20564b31bad991b4136a9815