Use package module to install distro packages
Consolidate distro package install tasks into a
single task using the package module and pass
the package list into the name instead of using
a with_items loop.

The minimum Ansible version is raised to 2.2 due to a
known bug [1] in Ansible's apt module which does not
update the cache properly if the cache update and the
install are combined in a single task.

[1] https://github.com/ansible/ansible-modules-core/issues/1497

Change-Id: I5d87e5e5d11bd1dcffb6a6187733f6d8e4e72fbf
(cherry picked from commit edd2093370ca6efb580a0864a8077ae201b29cd8)