Add OpenStack-Ansible metadata
We need to add openstack ansible information in the role
metadata to be able to track role maturity. With it,
we can create a role maturity table and take decisions about
role deprecations.

The role maturity state is "unmaintained", conformly to our
role maturity guidelines and the decision taken in our last
community meeting.

Change-Id: Ife751beac4b8eda1e90ca8aa00b9d718fa35a53f