Initial OSA zuul v3 role jobs
This patch implements an initial set of jobs intended to match
the current job execution method. It does not intend to improve
how the jobs are executed - only to replicate what is currently
in openstack-infra/openstack-zuul-jobs and provide the platform
to iterate on.

This also removes the duplicate openstack_hosts role from
the ansible-role-requirements.

Depends-On: https://review.openstack.org/556281
Change-Id: I4f4c9541d80d3c1cdaae7e9ab7f839aac92663b0