Add OpenStack-Ansible metadata
We need to add openstack ansible information in the role
metadata to be able to track role maturity. With it,
we can create a role maturity table and take decisions about
role deprecations.

The role maturity state is "incubated", conformly to our
role maturity guidelines. This role work has been started in
ocata and can therefore still have the incubated status in Pike.

Change-Id: Ia2ca7803115547931b452eee778afd02d3e54137