Make unstable functional test non-voting
Given that the functional build for this repo
rarely passes and the repo itself is only
maintained by those who use it, we can set the
functional test to non-voting in order to allow
general pattern changes to be implemented
without delay.

This does mean that reviewers applying functional
changes will have to be disciplined about checking
the results, but I think that is OK in this case.

Depends-On: https://review.openstack.org/556253
Change-Id: I1a236b88a8c80d82f162a92f6cdfeb9368b6eea7