Use monasca venv for influxdb setup
Instead of pip installing onto the host, we use the
influxdb in the monasca venv to do the influxdb database,
user, etc setup. This prevents interference between pip
installed packages and distribution packages, keeping the
line between host installed packages and venv installed
packages clean.

The venv is only deployed on the monasca_api servers, so
we change the group used for executing the tasks to the
right group for it to work.

Change-Id: I6e1f86a789f188dff611d522a3309a355f7cd8e8